Refactor joins step 1
